Port block and item logic to 1.21.11 API changes
- Update overridden block/item/entity methods to new 1.21.11 signatures and visibility
- Replace removed hooks with current equivalents for neighbor updates, shape ticks, and fall handling
- Fix related behavior updates (tool tags, data components, and codec/clone item handling)

+++ src/main/java/vectorwing/farmersdelight/common/block/RiceBaleBlock.java
@@ -21,7 +21,7 @@ public class RiceBaleBlock extends Block
 		this.registerDefaultState(this.getStateDefinition().any().setValue(FACING, Direction.UP));
 	}
 	@Override
-	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entityIn, float fallDistance) {
+	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entityIn, double fallDistance) {
 		entityIn.causeFallDamage(fallDistance, 0.2F, level.damageSources().fall());
 	}
 	@Override
