+++ src/main/java/vectorwing/farmersdelight/common/block/RichSoilFarmlandBlock.java
@@ -97,7 +97,7 @@ public class RichSoilFarmlandBlock extends FarmBlock
 		return !this.defaultBlockState().canSurvive(context.getLevel(), context.getClickedPos()) ? ModBlocks.RICH_SOIL.get().defaultBlockState() : super.getStateForPlacement(context);
 	}
 	@Override
-	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
 		entity.causeFallDamage(fallDistance, 1.0F, entity.damageSources().fall());
 	}
 }
